+++ nnet/src/nnet.cpp
@@ -1,7 +1,7 @@
 #include "nnet/nnet.hpp"
 
 namespace nnet {
-    std::vector<float> classify(const std::vector<uint8_t>& image) {
+    std::vector<double> classify(const std::vector<uint8_t>& image) {
         // Dummy implementation
         return {0.1, 0.05, 0.2, 0.05, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1};
     }
